pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "protectrag"
version = "0.6.0"
description = "Evals and observability for RAG: prompt-injection screening on stored documents"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [{ name = "ProtectRAG contributors" }]
keywords = ["rag", "llm", "security", "prompt-injection", "observability", "evals"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]
dependencies = []

[project.urls]
Homepage = "https://github.com/Ritvik777/ProtectRAG"
Repository = "https://github.com/Ritvik777/ProtectRAG"
Issues = "https://github.com/Ritvik777/ProtectRAG/issues"

[project.optional-dependencies]
llm = ["httpx>=0.25.0"]
otel = ["opentelemetry-api>=1.20.0", "opentelemetry-sdk>=1.20.0"]
langchain = ["langchain-core>=0.1.0"]
llamaindex = ["llama-index-core>=0.10.0"]
fastapi = ["fastapi>=0.100.0", "starlette>=0.27.0"]
redis = ["redis>=4.5.0"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23.0",
    "build>=1.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "twine>=5.0",
    "fakeredis>=2.20.0",
]

[tool.setuptools]
license-files = ["LICENSE"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
protectrag = ["py.typed", "data/*.json"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
